authorizations implemented

+++ src/api/notes/handler.js
@@ -13,8 +13,9 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload)
       const { title = 'untitled', body, tags } = request.payload
+      const { id: credentialId } = request.auth.credentials
 
-      const noteId = await this._service.addNote({ title, tags, body })
+      const noteId = await this._service.addNote({ title, tags, body, owner: credentialId })
       const response = h.response({
         status: 'success',
         message: 'Catatan berhasil ditambahkan',
@@ -45,7 +46,8 @@ class NotesHandler {
   }
 
   async getNotesHandler (request, h) {
-    const notes = await this._service.getNotes()
+    const { id: credentialId } = request.auth.credentials
+    const notes = await this._service.getNotes(credentialId)
     const response = h.response({
       status: 'success',
       data: { notes }
@@ -56,6 +58,9 @@ class NotesHandler {
   async getNoteByIdHandler (request, h) {
     try {
       const { id } = request.params
+      const { id: credentialId } = request.auth.credentials
+
+      await this._service.verifyNoteOwner(id, credentialId)
       const note = await this._service.getNoteById(id)
 
       return h.response({
@@ -87,7 +92,9 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload)
       const { id } = request.params
+      const { id: credentialId } = request.auth.credentials
 
+      await this._service.verifyNoteOwner(id, credentialId)
       await this._service.editNoteById(id, request.payload)
 
       return h.response({
@@ -118,7 +125,11 @@ class NotesHandler {
   async deleteNoteByIdHandler (request, h) {
     try {
       const { id } = request.params
+      const { id: credentialId } = request.auth.credentials
+
+      await this._service.verifyNoteOwner(id, credentialId)
       await this._service.deleteNoteById(id)
+
       return h.response({
         status: 'success',
         message: 'Catatan berhasil dihapus'
